% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genDataABN.R
\name{genDataABN}
\alias{genDataABN}
\title{Title}
\usage{
genDataABN(
  n = 100,
  p = 60,
  a = 6,
  b = 2,
  rho = 0.5,
  family = c("gaussian", "binomial"),
  signal = c("homogeneous", "heterogeneous"),
  noise = c("exchangeable", "autoregressive"),
  rho.noise = 0,
  beta,
  SNR = 1
)
}
\arguments{
\item{n}{holder}

\item{p}{holder}

\item{a}{holder}

\item{b}{holder}

\item{rho}{holder}

\item{family}{holder}

\item{signal}{holder}

\item{noise}{holder}

\item{rho.noise}{holder}

\item{beta}{holder}

\item{SNR}{holder}
}
\description{
Title
}
