% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quantiles.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Title}
\usage{
get_quantiles(quantile_functions, weights, len)
}
\arguments{
\item{quantile_functions}{holder}

\item{weights}{holder}

\item{len}{holder}
}
\description{
Title
}
